Fix automatic closure domain argument adaption (#370)
Previously the domain of a fencil closure was automatically set to
 ```unstructured_domain(named_range("horizontal", horizontal_start, horizontal_end)), named_range("K", vertical_start, vertical_end))```
regardless of what the horizontal dimension actually was. This lead to some problems with the temporary pass requires the correct name of the dimension. This PR fixes that by generating something like:
 ```unstructured_domain(named_range("Cell", horizontal_start, horizontal_end)), named_range("K", vertical_start, vertical_end))```
Additionally I generalized the function to support multiple closures, but kept the error message because I didn't test it.