Fix CI (0d field slice) (#383)
The latest commit on gt4py introduces a syntax feature: if you slice a field with full indexing, the result is still a field, a 0d field specifically. In order to use it as a scalar, the DSL program has to call .as_scalar(). Before this gt4py commit, the result of slicing would have been a scalar.

This PR adopts the new gt4py syntax feature. In order to do this, we bump the version of gt4py Spack package in stable CI job to 1.0.3.1.