Enable `optional` stencils and modules in Liskov (#358)
This PR allows to integrate optional stencils using Liskov. The GT4Py stencils are ignored if their module e.g. advection is not specified explicitly at compilation. The motivation for this feature is that the performance of some modules (e.g. advection and graupel) is not yet satisfying but keeping them in a separate branch is a lot of overhead.

More specifically this PR:

    adds a new argument parameter optional_module to the START STENCIL liskov directive to specify which optional module the stencil belongs to, e.g. optional_module=advection.
    adds a new option --enable_optional_stencil to the liskov CLI to enable the optional stencils and specify which module. e.g. --enable_optional_stencil=advection. Corresponding PR in icon-exclaim adds the changes in the icon_liskov preprocessor command in configure and CMakelist.txt.

---------

Co-authored-by: Nina Burgdorfer <Nina.Burgdorfer@meteoswiss.ch>
Co-authored-by: samkellerhals <kellerhalssamuel@gmail.com>
Co-authored-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Co-authored-by: Nicoletta Farabullini <41536517+nfarabullini@users.noreply.github.com>