Run icon4py and datatests on gpu (#640)
Enable driver runs and datatests on gpu. Only single-node run is tested
in this PR.

Major changes related to driver runs on gpu:
- A backend option is added into one of the click options
(`icon4py_driver_backend` or shortform `-b`). `xp` is decided explicitly
by calling the function in `data_allocation.py` that returns np if cpu
backends are used or cp if gpu backends are used.
Computation in `vertical.py` is changed to numpy and a backend argument
is added in `get_vct_a_and_vct_b` function to create `vct_a` and `vct_b`
fields with the correct backend.

- A new field_operator `_interpolate_to_half_levels_wp` is used in
`compute_virtual_potential_temperatures_and_pressure_gradient.py` to
avoid CUDA illegal memory access for an unknown reason.

- Because backend is added as an instance variable of `IconSavepoint`,
fields that are read from the serialized data are automatically
allocated on the device if gpu backend is used.

Major changes related to datatests on gpu:
- Add backend as an instance variable of `IconSavepoint` to create gt4py
fields on the device if running on GPU.

- Specify backend whenever allocation of gt4py fields is performed in
all datatests.

- Connectivities are allocated with `xp.asarray` when constructing
IconGrid in `grid_manager.py` and `serialbox.py`.

Note:
- test_advection.py fails on GPU. Further debugging is required.

---------

Co-authored-by: Hannes Vogt <vogt@hey.com>
Co-authored-by: Magdalena Luz <luzm@ethz.ch>