Add ltkeshs option, cleanup py2fgen and diffusion wrapper (#665)
Changes in the model:
- Adds the `ltkeshs` options to diffusion

Changes in the diffusion wrapper:
- The size arrays in `grid_init_diffusion` are extracted from their
fields and copied from cp to np. They should not live on the GPU in the
first place, which will be cleaned up in a next PR.
- The wrapper has to deal with optional fields: before xp.zeros where
passed from the bindings, however that was error-prone as we didn't have
all sizes available. After the refactoring of passing all sizes, this is
not even an option as the sizes are completely unknown.

Changes in py2fgen:
- Removes the `limited_area` command line flag, this is possible
since...
- Adds support for optional fields: if a parameter is annotated with
`Optional`, the Fortran parameter will be get the pointer attribute. If
the Fortran pointer is `associated` we pass the pointer via
`iso_c_bindings`, else we pass the `c_nullptr` and check for `ffi.NULL`
- For each Fortran array pass all its sizes. Before we passed a unique
size per dimension which might lead to subtle issues in case of wrong
typing. This is also a preparation to support plain arrays (not only
`gtx.Field`s).
- Change the default for `ICON4PY_PARALLEL` to `True`. For debugging we
keep the possibility to disable parallel mode.
- Introduced a new function `as_field()` to convert the ffi ptrs to
`gtx.Field`s