Update to GT4Py:icon4py_staging (follows GT4Py:main) (#596)
In this PR we switch to GT4Py:main plus the `concat_where` feature which
will get some more improvements before it will be available in main.

The latest gt4py PR for icon4py will be tracked in PR
https://github.com/GridTools/gt4py/pull/1928, from which we will create
tags.

Changes:
- Update to `gtx.Connectivity` from `NeighborTableOffsetProvider`
- Replace `where` by `concat_where` where semantically needed. There are
a few more places that will be fixed in a next PR.

---------

Co-authored-by: Hannes Vogt <hannes.vogt@cscs.ch>
Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>
Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>
Co-authored-by: Christoph Müller <Christoph.Mueller@meteoswiss.ch>
Co-authored-by: Magdalena Luz <luzm@ethz.ch>