Bencher: feature branches (#669)
Compare the performance results of the current branch (PR's branch) with
Bencher's baseline (main branch)
(https://github.com/C2SM/icon4py/pull/666 &
https://github.com/C2SM/icon4py/pull/695)

This pipeline compares the performance of the feature branch with the
baseline found in Bencher cloud (not the self-hosted one -for simplicity
and not any other reason-)

Before merging any feature branch to main, the author should run
`cscs-ci run benchmark-bencher` (before it was just `cscs-ci run
benchmark`). If there is performance regression, the pipeline fails and
the PR author can inspect which tests are failing.

---------

Co-authored-by: Ioannis Magkanaris <ioannis.magkanaris@cscs.ch>