add test level marker (#722)
Many function in for the static field computations in `icon4py` are
tested twice:
1. in isolation, with input and output from serialized data (eg as
`unit` tests)
2. integrated into the factories:
[interpolation_factory](model/common/src/icon4py/model/common/interpolation/interpolation_factory.py)
and
[metrics_factory](model/common/src/icon4py/model/common/metrics/metrics_factory.py)
with _computed_ inputs and verified against serialized outputs.

The unit level tests are useful for fast feedback and refactoring these
function but as the demand on the CI is high it is not necessary to run
both version of them all the time. Hence, this PR

- marks tests with either of the level (where both level exist)
- a pytest option `--level=integration|unit|any` to selectively run
either of the levels
- runs only `integration` level tests in the default CI action

**Additional changes**
- collect some related functions into same module.