Run pre-commit with `--all-files` flag in CI (#743)
`pre-commit` was not doing any checks in CI, because it was run without
`--all-files`, which normally means running on all changed files about
to be committed. Since there are no staged changes in the CI run, no
files were checked.

This silences mypy type warnings and formatting errors that snuck in
while the checks were disabled. This also updates the actions used in
the QA workflow and uses `setup-python`, `setup-uv`, and `pre-commit`
actions instead of manually setting up the tools. The `.python-version`
file is no longer ignored in `.gitignore` as it's actually checked into
the repo and used by CI and environments.