Refactorings in IconGrid (#755)
**(Refactor) refactorings related to `IconGrid` and `dimensions.py`:**
- add tests for the correct setting of `skip_value` in icon4py
dimensions.
- type fixes: use the correct type for `gt4py.Dimension` implementation.
- the refinement control fields are added to `IconGrid`: that is the
best bet for them as they encode grid topology information (similar (but
more accurate for limited area grids) to `start_index` and `end_index`.
- `limited_area = [True|False]` is now determined from the `refinement
control`fields in grid file in `grid_manager.py`. The argument injection
to the `GridConfig` is kept to allow passing it in `py2fgen`

- cleanup of inconsistent naming: 
- the legacy `offset_provider` is removed from `BaseGrid`
implementations.
- functionality dealing with `gt4py.Connectivity` now use
`connectivity`.
- simple `ndarray` s dealing with neighbor relations use
`neighbor_tables`
- GT4Py's `as_connectivity` *copies* the neighbor table. Such that the
grid implementation contain the neighbor tables twice: once in the
original `_neighbor_table` dict and inside the `connectivities` dict.
Once we start maninpulating this can lead to inconsistent tables being
used. Hence `BaseGrid.neighbor_tables` property is added that accesses
the neighbor tables from the `connectivity` dict. The idea is to
eventually hide (or remove) the original `_neighbor_tables` completely.
- builder functions for the `IconGrid` have been renamed to from `with_`
to `set_` as the manipulate in place and do not return a new instance.