add download uri for "new grid files" (#766)
This PR collects the information for validation experiments (serialized
data sets) and grids used in testing in benchmarking of `icon4py` in a
new structure and central place (`model/testing/definitions.py`) It also
adds data for
additional grids to be used for benchmarking and (future) validation.


The new structure is not used yet in the code, this is left to
subsequent PRs.

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>