Replace skip values (#761)
Add a function that removes remaining skip values from the neighbor
tables. This replaces [PR 728](https://github.com/C2SM/icon4py/pull/728)

- Replacing the skip values is configurable by a flag in the
`GridConfig`.
- when `GridConfig.keep_skip_values = False` the skip values in the
neighbortables of the `Connectivity` will be replaced by the max index
in the neighborhood of an entity. This is done for all except for the
pentagon points in a global grid.

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>