Remove unused stencil init_two_cell_kdim_fields_index_with_zero_vp (#777)
The stencil test for `init_two_cell_kdim_fields_index_with_zero_vp` was
failing with a crash on the dace backend. The cause was wrong input data
(`k2=grid.num_levels`) that caused an out-of-bound memory access.

It was also found that this stencil is not used anywhere. Instead of
fixing the test, this PR is removing the stencil.