RBF interpolation stencils improvement (#709)
Adds computation of RBF interpolation coefficient fields for cells,
edges, and vertices. Adds a few missing cartesian coordinate fields used
as inputs for the computation. The R04B09 vertex and cell fields are
quite far off from the reference fields (~1e-2) but this seems to be
inherent to the method.

This does not port the gradient RBF field.

This is using numpy/cupy for the computation and is mostly vectorized.
The cholesky decomposition/solve is always done on CPU, and is not
vectorized/parallelized. The RBF matrices are small (n=4-9) and don't
offer enough parallelism. I haven't made any attempts at parallelizing
this or profiling in general, so this may be suboptimal, but I would
leave potential optimizations for a follow-up, if needed.

---------

Co-authored-by: Mikael Simberg <mikael.simberg@iki.fi>
Co-authored-by: Mikael Simberg <simberg@cscs.ch>