Bump GT4Py version (includes dace fix) (#785)
The new gt4py version `icon4py_staging_20250704` contains fixes for the
dace backend. However, the dace backend is not yet enabled in CI, this
will be done in a separate PR.

Note that this PR disables the tests on orchestrated diffusion granule
because dace orchestration is currently not compatible with precompiled
programs.