muphys final version  (#786)
I addressed most of the issues raised by Nikki's review. One outstanding
issue is to document function arguments with triple quotes. I'm not sure
I want this everywhere, since it will substantially add to the length of
the code (we have lots of arguments...).

All 54 unit tests pass with embedded and gtfn_cpu on Santis. The graupel
application only runs with embedded, and currently is not working with
gtfn or dace.

---------

Co-authored-by: Will Sawyer <wsawyer@cscs.ch>
Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>
Co-authored-by: Nicoletta Farabullini <41536517+nfarabullini@users.noreply.github.com>
Co-authored-by: Magdalena <luzm@ethz.ch>