fix bug in computation of dwdz for 3d divergence damping  (#793)
`starting_vertical_index_for_3d_divdamp` is wrong when `divdamp_type =
32`.
`starting_vertical_index_for_3d_divdamp` is used as a manual
optimization in the original ICON. It is a k starting index for
computation of `dwdz_at_cells_on_model_levels` (declared as `z_dwdz_dd`
in ICON) which is used later in the computation of 3D divergence
damping.
We have decided to remove this index in icon4py. The purpose of this PR
is to
(1) remove `starting_vertical_index_for_3d_divdamp` and
(2) add verification for `dwdz_at_cells_on_model_levels` in
`test_vertically_implicit_solver_at_predictor_step` that was missing
before.