refactor: refactor test infrastructure (#804)
This PR changes the structure of the tests in the repo and adds utility
scripts to help keeping it sane. The new structure is described in the
Coding Conventions document. This is just the first step of a larger
effort to refactor and clean up the tests infrastructure and only deals
with the folder structure.

---------

Co-authored-by: Hannes Vogt <vogt@hey.com>