Refactor global grid params for better torus support (#829)
This is the first of n PRs attempting to structure torus support better
as part of https://hackmd.io/SNLrlU-eR_KsIWjQGigzTg?view.

This first PR only does a small refactoring:
- Add a few helper types for better managing the root, level, geometry
type parameters.
- Uses the above types when guessing the grid parameters from file names
(I've renamed the function `guess_grid_type` to signal that it's a guess
based on file name rather than a structured way of getting the
information from the grid files).
- Adds better validation for torus/sphere parameters.
- To help with the above, it adds a `try_attribute` member function to
`GridFile` to read optional attributes (in this case we want to read
`geometry_type` if it exists, and ignore it if it's not in the grid
file).
- Fixes a few flyby docstrings and comments.
- Adds a few more type hints to the global grid parameters helper
functions.