Move nflat gradp to metrics (#856)
Moves `nflat_gradp` out of the `VerticalGrid` placing it in the
`NonHydroMetricState`, because it can only be computed together with
`flat_idx` from the metrics fields.

Accordingly, `nflat_gradp` has been removed in in the `diffusion` since
it is not used, then in `dycore` it has been moved to the metrics field.
See also [interface change in
icon-exclaim](https://github.com/C2SM/icon-exclaim/pull/372)

Additional change: delete an old unused dycore stencil (51).