Clean up grid params construction (#847)
Part 2 of N for better torus support. This requires
https://github.com/C2SM/icon4py/pull/829.

This only removes the `from_mean_cell_area` constructor from
`GridParams` as a cleanup. I realized it doesn't add much value since it
doesn't do any extra logic or validation, all it did was provide a
convenient way to pass the `mean_cell_area` as the first argument. I've
made the `GridParams` constructor take keyword-only parameters so it's
easy to pass only the `mean_cell_area` with the regular constructor.