Scalar values from factory (#871)
Adds the ability to produce scalar values from the factories. 
We need this in the metrics factory for the computation of `nflat_gradp`
for the dynamical core and will need it as well for the computation of
some mean properties for the grid (`mean_cell_area`...) as discussed in
[PR-848](https://github.com/C2SM/icon4py/pull/848)

(Additional fixes): fixes a lot of typing issues in all three factory
implementations.

---------

Co-authored-by: yiluchen1066 <67939569+yiluchen1066@users.noreply.github.com>