Microphysics graupel (#537)
Port of the ICON microphysics graupel scheme. The graupel scheme code
has been refactored in favor of a more readable style. Currently, we
verify it by using the serialized data from Weisman-Klemp experiment
performed on the torus grid.

The graupel scheme is split into three parts:
1. the main algorithm for computing the temperature and tracer
tendencies due to cloud microphyisal processes. It is wrapped in a scan
operator because the original graupel scheme contains implicit
computation for the sedimentation fluxes. A number of local variables
are also output for determination of rain, snow, graupel, and ice
sedimentation fluxes in the following steps.
2. computation of rain, snow, graupel, and ice sedimentation fluxes
above the ground.
3. computation of rain, snow, graupel, and ice sedimentation fluxes on
the surface.
All the pre-computed coefficients in `gscp_data.f90` are now stored in
the `FrozenNameSpace` class `MicrophysicsConstants`. They can be
directly used in the scan operator.

---------

Co-authored-by: Magdalena Luz <luzm@ethz.ch>