Better torus support for mean cell area etc. (#848)
Part 3 of N for better torus support.

Adds:
- `mean_cell_area`
- `mean_dual_cell_area`
- `mean_edge_length`
- `mean_dual_edge_length`

and their computation to `GlobalGridParams`. These are read from the
grid file, if available. If they're not available, they're computed from
the corresponding fields. As a fallback of the fallback, the
`mean_cell_area` is computed from the global number of cells and the
area of the sphere. However, I'm considering removing this fallback as
it should always be possible to compute it from the fields.

As far as I can tell, `mean_dual_cell_area` and `mean_edge_length` are
currently not used for anything but I'm adding them for completeness
because:
- `mean_cell_area` is required to compute `characteristic_length` which
is used for RBF interpolation on spheres, and
- `mean_dual_edge_length` is used for RBF interpolation on toruses (this
will be used in #853).