fix: cache ghex pattern on buffer and dimension (#945)
In #915 I claimed that the dimension does not have be part of the key as
I thought the buffer uniquely identifies the exchange pattern. However
if a field gets de-allocated and the same buffer gets re-allocated for a
field with different dimension (which can only happen if all dimensions
have the same length (`nproma`)) then the caching is broken.

This case can only happen in blueline verification mode (where fields
are allocated/deallocated in the timeloop).