Fix e flx average or e2c2c (#944)
(FIX): the computation of e_flx_avg from the factory did not verify and
was not tested for the factory computation, the simple function test
against serialized data verfied.
We could track it down to being somehow dependent on the local ordering
of the diamond connectivities being exactly in the order that icon uses.

For simplicity we fixed the ordering of e2c2e in the grid manager.